Complete code reorganization
