+++ src/request/timeline.rs
@@ -1,11 +1,14 @@
+pub use self::err::TimelineErr;
+pub use self::inner::{Content, Reach, Scope, Stream, UserData};
 use super::query::Query;
-use crate::err::TimelineErr;
-use crate::event::Id;
 
-use hashbrown::HashSet;
 use lru::LruCache;
-use std::convert::TryFrom;
 use warp::reject::Rejection;
+
+mod err;
+mod inner;
+
+type Result<T> = std::result::Result<T, TimelineErr>;
 
 #[derive(Clone, Debug, Copy, Eq, Hash, PartialEq)]
 pub struct Timeline(pub Stream, pub Reach, pub Content);
@@ -15,7 +18,7 @@ impl Timeline {
         Self(Stream::Unset, Reach::Local, Content::Notification)
     }
 
-    pub fn to_redis_raw_timeline(&self, hashtag: Option<&String>) -> Result<String, TimelineErr> {
+    pub fn to_redis_raw_timeline(&self, hashtag: Option<&String>) -> Result<String> {
         use {Content::*, Reach::*, Stream::*};
         Ok(match self {
             Timeline(Public, Federated, All) => "timeline:public".into(),
@@ -39,34 +42,27 @@ impl Timeline {
         })
     }
 
-    pub fn from_redis_text(
-        timeline: &str,
-        cache: &mut LruCache<String, i64>,
-    ) -> Result<Self, TimelineErr> {
-        // TODO -- can a combinator shorten this?
-        let mut id_from_tag = |tag: &str| match cache.get(&tag.to_string()) {
-            Some(id) => Ok(*id),
-            None => Err(TimelineErr::InvalidInput), // TODO more specific
-        };
+    pub fn from_redis_text(timeline: &str, cache: &mut LruCache<String, i64>) -> Result<Self> {
+        use {Content::*, Reach::*, Stream::*, TimelineErr::*};
+        let mut tag_id = |t: &str| cache.get(&t.to_string()).map_or(Err(BadTag), |id| Ok(*id));
 
-        use {Content::*, Reach::*, Stream::*};
         Ok(match &timeline.split(':').collect::<Vec<&str>>()[..] {
             ["public"] => Timeline(Public, Federated, All),
             ["public", "local"] => Timeline(Public, Local, All),
             ["public", "media"] => Timeline(Public, Federated, Media),
             ["public", "local", "media"] => Timeline(Public, Local, Media),
-            ["hashtag", tag] => Timeline(Hashtag(id_from_tag(tag)?), Federated, All),
-            ["hashtag", tag, "local"] => Timeline(Hashtag(id_from_tag(tag)?), Local, All),
+            ["hashtag", tag] => Timeline(Hashtag(tag_id(tag)?), Federated, All),
+            ["hashtag", tag, "local"] => Timeline(Hashtag(tag_id(tag)?), Local, All),
             [id] => Timeline(User(id.parse()?), Federated, All),
             [id, "notification"] => Timeline(User(id.parse()?), Federated, Notification),
             ["list", id] => Timeline(List(id.parse()?), Federated, All),
             ["direct", id] => Timeline(Direct(id.parse()?), Federated, All),
             // Other endpoints don't exist:
-            [..] => Err(TimelineErr::InvalidInput)?,
+            [..] => Err(InvalidInput)?,
         })
     }
 
-    pub fn from_query_and_user(q: &Query, user: &UserData) -> Result<Self, Rejection> {
+    pub fn from_query_and_user(q: &Query, user: &UserData) -> std::result::Result<Self, Rejection> {
         use {warp::reject::custom, Content::*, Reach::*, Scope::*, Stream::*};
 
         Ok(match q.stream.as_ref() {
@@ -104,71 +100,5 @@ impl Timeline {
                 Err(custom("Error: Nonexistent endpoint"))?
             }
         })
-    }
-}
-
-#[derive(Clone, Debug, Copy, Eq, Hash, PartialEq)]
-pub enum Stream {
-    User(Id),
-    // TODO consider whether List, Direct, and Hashtag should all be `id::Id`s
-    List(i64),
-    Direct(i64),
-    Hashtag(i64),
-    Public,
-    Unset,
-}
-
-#[derive(Clone, Debug, Copy, Eq, Hash, PartialEq)]
-pub enum Reach {
-    Local,
-    Federated,
-}
-
-#[derive(Clone, Debug, Copy, Eq, Hash, PartialEq)]
-pub enum Content {
-    All,
-    Media,
-    Notification,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub enum Scope {
-    Read,
-    Statuses,
-    Notifications,
-    Lists,
-}
-
-impl TryFrom<&str> for Scope {
-    type Error = TimelineErr;
-
-    fn try_from(s: &str) -> Result<Self, Self::Error> {
-        match s {
-            "read" => Ok(Scope::Read),
-            "read:statuses" => Ok(Scope::Statuses),
-            "read:notifications" => Ok(Scope::Notifications),
-            "read:lists" => Ok(Scope::Lists),
-            "write" | "follow" => Err(TimelineErr::InvalidInput), // ignore write scopes
-            unexpected => {
-                log::warn!("Ignoring unknown scope `{}`", unexpected);
-                Err(TimelineErr::InvalidInput)
-            }
-        }
-    }
-}
-
-pub struct UserData {
-    pub id: Id,
-    pub allowed_langs: HashSet<String>,
-    pub scopes: HashSet<Scope>,
-}
-
-impl UserData {
-    pub fn public() -> Self {
-        Self {
-            id: Id(-1),
-            allowed_langs: HashSet::new(),
-            scopes: HashSet::new(),
-        }
     }
 }
