+++ src/response/redis/connection.rs
@@ -1,13 +1,13 @@
 mod err;
 pub use err::RedisConnErr;
 
-use super::msg::{RedisMsg, RedisParseErr, RedisParseOutput};
-use super::ManagerErr;
+use super::msg::{RedisParseErr, RedisParseOutput};
+use super::{ManagerErr, RedisCmd};
 use crate::config::Redis;
 use crate::event::Event;
 use crate::request::{Stream, Timeline};
 
-use futures::Async;
+use futures::{Async, Poll};
 use lru::LruCache;
 use std::convert::{TryFrom, TryInto};
 use std::io::{Read, Write};
@@ -16,7 +16,6 @@ use std::str;
 use std::time::Duration;
 
 type Result<T> = std::result::Result<T, RedisConnErr>;
-type Poll = futures::Poll<Option<(Timeline, Event)>, ManagerErr>;
 
 #[derive(Debug)]
 pub struct RedisConn {
@@ -48,7 +47,7 @@ impl RedisConn {
         Ok(redis_conn)
     }
 
-    pub fn poll_redis(&mut self) -> Poll {
+    pub fn poll_redis(&mut self) -> Poll<Option<(Timeline, Event)>, ManagerErr> {
         let mut size = 100; // large enough to handle subscribe/unsubscribe notice
         let (mut buffer, mut first_read) = (vec![0u8; size], true);
         loop {
@@ -68,6 +67,7 @@ impl RedisConn {
             return Ok(Async::NotReady);
         }
 
+        // at this point, we have the raw bytes; now, parse what we can and leave the remainder
         let input = self.redis_input.clone();
         self.redis_input.clear();
 
@@ -83,11 +83,15 @@ impl RedisConn {
             Ok(Msg(msg)) => match &self.redis_namespace {
                 Some(ns) if msg.timeline_txt.starts_with(&format!("{}:timeline:", ns)) => {
                     let trimmed_tl = &msg.timeline_txt[ns.len() + ":timeline:".len()..];
-                    (self.into_tl_event(trimmed_tl, &msg)?, msg.leftover_input)
+                    let tl = Timeline::from_redis_text(trimmed_tl, &mut self.tag_id_cache)?;
+                    let event = msg.event_txt.try_into()?;
+                    (Ok(Ready(Some((tl, event)))), (msg.leftover_input))
                 }
                 None => {
                     let trimmed_tl = &msg.timeline_txt["timeline:".len()..];
-                    (self.into_tl_event(trimmed_tl, &msg)?, msg.leftover_input)
+                    let tl = Timeline::from_redis_text(trimmed_tl, &mut self.tag_id_cache)?;
+                    let event = msg.event_txt.try_into()?;
+                    (Ok(Ready(Some((tl, event)))), (msg.leftover_input))
                 }
                 Some(_non_matching_namespace) => (Ok(Ready(None)), msg.leftover_input),
             },
@@ -98,12 +102,6 @@ impl RedisConn {
         self.redis_input.extend_from_slice(leftover.as_bytes());
         self.redis_input.extend_from_slice(invalid_bytes);
         res
-    }
-
-    fn into_tl_event<'a>(&mut self, tl: &'a str, msg: &'a RedisMsg) -> Result<Poll> {
-        let tl = Timeline::from_redis_text(tl, &mut self.tag_id_cache)?;
-        let event = msg.event_txt.try_into().expect("TODO");
-        Ok(Ok(Async::Ready(Some((tl, event)))))
     }
 
     pub fn update_cache(&mut self, hashtag: String, id: i64) {
@@ -135,10 +133,11 @@ impl RedisConn {
             .map_err(|e| RedisConnErr::with_addr(&addr, e))?;
         Ok(conn)
     }
+
     fn auth_connection(conn: &mut TcpStream, addr: &str, pass: &str) -> Result<()> {
         conn.write_all(&format!("*2\r\n$4\r\nauth\r\n${}\r\n{}\r\n", pass.len(), pass).as_bytes())
             .map_err(|e| RedisConnErr::with_addr(&addr, e))?;
-        let mut buffer = vec![0u8; 5];
+        let mut buffer = vec![0_u8; 5];
         conn.read_exact(&mut buffer)
             .map_err(|e| RedisConnErr::with_addr(&addr, e))?;
         if String::from_utf8_lossy(&buffer) != "+OK\r\n" {
@@ -150,7 +149,7 @@ impl RedisConn {
     fn validate_connection(conn: &mut TcpStream, addr: &str) -> Result<()> {
         conn.write_all(b"PING\r\n")
             .map_err(|e| RedisConnErr::with_addr(&addr, e))?;
-        let mut buffer = vec![0u8; 7];
+        let mut buffer = vec![0_u8; 7];
         conn.read_exact(&mut buffer)
             .map_err(|e| RedisConnErr::with_addr(&addr, e))?;
         let reply = String::from_utf8_lossy(&buffer);
@@ -159,26 +158,6 @@ impl RedisConn {
             "-NOAUTH" => Err(RedisConnErr::MissingPassword),
             "HTTP/1." => Err(RedisConnErr::NotRedis(addr.to_string())),
             _ => Err(RedisConnErr::InvalidRedisReply(reply.to_string())),
-        }
-    }
-}
-
-pub enum RedisCmd {
-    Subscribe,
-    Unsubscribe,
-}
-
-impl RedisCmd {
-    pub fn into_sendable(&self, tl: &String) -> (Vec<u8>, Vec<u8>) {
-        match self {
-            RedisCmd::Subscribe => (
-                format!("*2\r\n$9\r\nsubscribe\r\n${}\r\n{}\r\n", tl.len(), tl).into_bytes(),
-                format!("*3\r\n$3\r\nSET\r\n${}\r\n{}\r\n$1\r\n1\r\n", tl.len(), tl).into_bytes(),
-            ),
-            RedisCmd::Unsubscribe => (
-                format!("*2\r\n$11\r\nunsubscribe\r\n${}\r\n{}\r\n", tl.len(), tl).into_bytes(),
-                format!("*3\r\n$3\r\nSET\r\n${}\r\n{}\r\n$1\r\n0\r\n", tl.len(), tl).into_bytes(),
-            ),
         }
     }
 }
