+++ src/response/redis/connection/err.rs
@@ -1,4 +1,4 @@
-use crate::err::TimelineErr;
+use crate::request::TimelineErr;
 use std::fmt;
 
 #[derive(Debug)]
